// src/common/constants/enums.ts
export enum Icon {
    WARNING = "material-symbols/warning-rounded",
    FOLDER_QUESTION = "mdi/folder-question",
    FOLDER = "noto/file-folder", // 📁
    CLIPBOARD = "noto/clipboard", // 📋
    CHART = "noto/bar-chart", // 📊
}

export enum Urls {
    MEDIA_SERVER = "https://media.githubusercontent.com/media/YamanquiChacala/GeneradorDeReportes/refs/heads/main/",
}

export enum Templates {
    HTML_TO_PDF_TEMPLATE = "PdfPrintTemplate",
}

export enum Colors {
    LOGO_OSCURO = "#34666A",
    LOGO_CLARO = "#159A5E",
    ORANGE = "#EA4335",
    LANGUAGE = "#c9daf8",
    SCIENCE = "#fce5cd",
    NATURE = "#d9ead3",
    HUMANITIES = "#ead1dc",
}